Added confirmation question for deleting database

+++ make.sh
@@ -6,9 +6,12 @@ if [ ! -d "$BACKUP" ]; then
 mkdir $HOME/$BACKUP
 fi
 
+read -p "Do you want to delete your current database and start with a clean, new empty one?" -n 1 -r
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
 # Clean out the tangerine DB of any existing assessments or results
 curl -H "Content-Type: application/json" -X DELETE http://tangerine:tangytangerine@localhost:5984/tangerine; curl -H "Content-Type: application/json" -X PUT http://tangerine:tangytangerine@localhost:5984/tangerine; cd ../app; couchapp push; cd -
-
+fi
 
 echo
 echo "********"
